pyproject.toml
[project]
name = "gsw-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Shreyas Rajesh", email = "shreyasrajesh38@g.ucla.edu" }
]
requires-python = ">=3.10"
dependencies = [
    "bespokelabs-curator>=0.1.25",
    "editables>=0.5",
    "faiss-gpu>=1.7.2",
    "flash-attn>=2.8.3",
    "google-genai>=1.46.0",
    "hatchling>=1.28.0",
    "hdbscan>=0.8.40",
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "langchain-voyageai>=0.1.6",
    "numpy==1.26.4",
    "omegaconf>=2.3.0",
    "openai>=1.82.0",
    "pandas>=2.2.2",
    "peft>=0.17.1",
    "plotly>=6.5.2",
    "pyarrow>=21.0.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "rank-bm25>=0.2.2",
    "rich>=13.9.4",
    "scikit-learn>=1.6.1",
    "streamlit>=1.50.0",
    "together>=1.5.26",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
    "trl>=0.25.0",
    "unsloth>=2025.9.9",
    "verl>=0.7.0",
    "vllm>=0.10.0",
    "wandb>=0.22.3",
]

[project.scripts]
gsw-memory = "gsw_memory:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.11",
]
